+++ plugins/SkyCultureMaker/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.1
 
+* New Feature: Option to automatically merge lines into polylines on export
 * Fix: Drawn constellations are not reset after "Export and Exit"
 * Fix: "Export and Exit" exits even when export failed, deleting all unsaved progress
 * Fix: Stop drawing line key (Double Right Click) only works when Star/DSO is selected